+++ lib/src/operation/network_wrapper.dart
@@ -1,38 +1,41 @@
 part of '../network_manager.dart';
 
 extension _CoreServiceWrapperExtension on NetworkManager {
-  void _addNetworkInterceptors(InterceptorsWrapper? interceptor) {
+  void _addNetworkInterceptors(dio.Interceptor? interceptor) {
     if (interceptor != null) interceptors.add(interceptor);
     interceptors.add(_onErrorWrapper());
   }
 
-  InterceptorsWrapper _onErrorWrapper() {
-    return InterceptorsWrapper(
-      onError: (e, handler) async {
+  QueuedInterceptorsWrapper _onErrorWrapper() {
+    return QueuedInterceptorsWrapper(
+      onError: (DioError e, ErrorInterceptorHandler handler) async {
         final errorResponse = e.response;
         if (errorResponse == null) {
         } else {
-          if (errorResponse.statusCode == HttpStatus.unauthorized && onRefreshToken != null) {
-            if (_retryCount < _maxCount) {
-              _retryCount++;
-              interceptors.responseLock.lock();
-              interceptors.requestLock.lock();
-              final error = await onRefreshToken!(e, NetworkManager(options: options));
-              interceptors.responseLock.unlock();
-              interceptors.requestLock.unlock();
-              final requestModel = error.requestOptions;
+          if (errorResponse.statusCode == HttpStatus.unauthorized &&
+              onRefreshToken != null) {
+            final error =
+                await onRefreshToken!(e, NetworkManager(options: options));
+            final requestModel = error.requestOptions;
 
-              final response = await request(
-                requestModel.path,
-                queryParameters: requestModel.queryParameters,
-                data: requestModel.data,
-                options: Options(method: requestModel.method, headers: requestModel.headers),
+            try {
+              final response = await retry(
+                () => dio.Dio(options).request(
+                  requestModel.path,
+                  queryParameters: requestModel.queryParameters,
+                  data: requestModel.data,
+                  options: Options(
+                      method: requestModel.method,
+                      headers: requestModel.headers),
+                ),
+                maxAttempts: _maxCount,
+                retryIf: (e) {
+                  return e is TimeoutException || e is DioError;
+                },
               );
-
               return handler.resolve(response);
-            } else {
-              if (onRefreshFail != null) onRefreshFail!();
-              _retryCount = 0;
+            } catch (_) {
+              return handler.next(e);
             }
           }
         }